move names functions into separate file
