+++ src/utils/combo/index.ts
@@ -1,3 +1,2 @@
 export * from "./criteria";
 export * from "./filter";
-export * from "./matchNames";
